+++ listings/templates/create_listing.html
@@ -156,7 +156,7 @@ document.addEventListener("DOMContentLoaded", function() {
     // This function calls your Django endpoint (which returns JSON data)
     // and fills in both the suggested price field and the detailed table.
     function fetchSuggestedPrice(isbn) {
-        fetch("{% url 'listings:fetch_ebay_price' %}?isbn=" + isbn)
+        fetch("{% url 'pricing_engine:get_suggestions' %}?isbn=" + isbn)
             .then(response => response.json())
             .then(data => {
                 console.log("Fetched data:", data);
@@ -180,7 +180,7 @@ document.addEventListener("DOMContentLoaded", function() {
                 }
             })
             .catch(error => {
-                console.error("Error fetching price:", error);
+                console.error("Error fetching price suggestions:", error);
                 document.getElementById("suggested-price").value = "N/A";
                 document.getElementById("id_price").value = "";
             });
