+++ pricing_engine/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class PricingEngineConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "pricing_engine"
