+++ pricing_engine/urls.py
@@ -1,0 +1,8 @@
+from django.urls import path
+from . import views
+
+app_name = 'pricing_engine'
+
+urlpatterns = [
+    path("suggest/", views.EbayPriceScraperView.as_view(), name="get_suggestions"),
+]
