+++ textbookxchange/urls.py
@@ -24,6 +24,7 @@ urlpatterns = [
     path("admin/", admin.site.urls),
     path("", include("core.urls")),
     path("listings/", include("listings.urls", namespace='listings')),
+    path("pricing/", include("pricing_engine.urls", namespace='pricing_engine')),
 ]
 if settings.DEBUG:
     urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
